chore: package 이름 대문자에서 소문자로 변경
