+++ src/main/java/our/yurivongella/instagramclone/domain/post/PostLike.java
@@ -1,7 +1,4 @@
-package our.yurivongella.instagramclone.domain.Comment;
-
-import java.util.ArrayList;
-import java.util.List;
+package our.yurivongella.instagramclone.domain.post;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
@@ -11,37 +8,30 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
-import javax.persistence.OneToMany;
 import javax.persistence.Table;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import our.yurivongella.instagramclone.domain.BaseEntity;
-import our.yurivongella.instagramclone.domain.Post.Post;
-import our.yurivongella.instagramclone.domain.Users.Users;
+import our.yurivongella.instagramclone.domain.users.Users;
 
-@Table(name = "comment")
 @Getter
 @Entity
 @NoArgsConstructor
-public class Comment extends BaseEntity {
+@Table(name = "post_like")
+public class PostLike extends BaseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "comment_id")
+    @Column(name = "post_like_id")
     private Long id;
-
-    @Column(name = "content", columnDefinition = "text")
-    private String content;
-
-    @JoinColumn(name = "user_id")
-    @ManyToOne(fetch = FetchType.LAZY)
-    private Users user;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "post_id")
     private Post post;
 
-    @OneToMany(mappedBy = "comment")
-    private List<CommentLike> commentLikes = new ArrayList<>();
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "user_id")
+    private Users user;
+
 
 }
