+++ src/main/java/our/yurivongella/instagramclone/domain/post/PostLikeRepository.java
@@ -1,4 +1,4 @@
-package our.yurivongella.instagramclone.domain.Post;
+package our.yurivongella.instagramclone.domain.post;
 
 import org.springframework.data.jpa.repository.JpaRepository;
