+++ src/main/java/our/yurivongella/instagramclone/domain/users/UsersRepository.java
@@ -1,4 +1,4 @@
-package our.yurivongella.instagramclone.domain.Users;
+package our.yurivongella.instagramclone.domain.users;
 
 import org.springframework.data.jpa.repository.JpaRepository;
